+++ Client.h
@@ -4,12 +4,12 @@
 class Client {
 	private:
 		int h;				// N° du client
-		int dist;			// Distance entrepôt/client (en unité de temps)
-		float cout_u;			// Coût de stockage unitaire chez le client h
+		float dist;			// Distance entrepôt/client (en unité de temps)
+		double cout_u;			// Coût de stockage unitaire chez le client h
 		bool fini;			// Est-ce que le client a fini d'être livré ?
 
 	public:
-		Client(int, float, int);
+		Client(int, double, float);
 		float getCoutUnitaireStockage();
 		int getDist();
 		int getNum();
